Move market notify deal signifiers to better place
